+++ hw_printer_network/controllers/hw_printer_network_controller.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+# Copyright 2017-2018 Dinar Gabbasov <https://it-projects.info/team/GabbasovDinar>
+# Copyright 2018 Tom Blauwendraat <tom@sunflowerweb.nl>
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
+
 from openerp import http
 import logging
 import time
@@ -123,7 +127,9 @@ class EscposNetworkDriver(EscposDriver):
                             _logger.info('Done printing XML receipt on printer %s', network_printer_ip)
                         else:
                             _logger.error('xml_receipt: printer offline!')
-                            # TODO: do something with the missed order?
+                            # add a missed order to queue
+                            time.sleep(3)
+                            self.queue.put((timestamp, task, data))
                 elif task == 'printstatus':
                     pass
                 elif task == 'status':