+++ pos_restaurant_base/static/lib/js/models.js
@@ -8,6 +8,23 @@ odoo.define('pos_restaurant_base.models', function (require) {
 
     var QWeb = core.qweb;
 
+    var _super_orderline = models.Orderline.prototype;
+    models.Orderline = models.Orderline.extend({
+        set_dirty: function(dirty) {
+            //  DIFFERENCES FROM ORIGINAL:
+            // * check mp_dirty to avoid repeated orderline rendering
+            //   (https://github.com/odoo/odoo/pull/23266)
+            //
+            // * using orderline_change_line function instead trigger
+            //   allows you to avoid unnecessary multiple calls of the same functions
+            if (this.mp_dirty !== dirty) {
+                this.mp_dirty = dirty;
+                if (this.pos.gui.screen_instances.products) {
+                    this.pos.gui.screen_instances.products.order_widget.orderline_change_line(this);
+                }
+            }
+        }
+    });
 
     var _super_order = models.Order.prototype;
     models.Order = models.Order.extend({
@@ -165,9 +182,21 @@ odoo.define('pos_restaurant_base.models', function (require) {
             }
         },
         print_order_receipt: function(printer, changes) {
+            var self = this;
+            function delay(ms) {
+                var d = $.Deferred();
+                setTimeout(function(){
+                    d.resolve();
+                }, ms);
+                return d.promise();
+            }
+            var q = $.when();
             if ( changes['new'].length > 0 || changes['cancelled'].length > 0){
-                var receipt = QWeb.render('OrderChangeReceipt',{changes:changes, widget:this});
-                printer.print(receipt);
+                q = q.then(function(){
+                    var receipt = QWeb.render('OrderChangeReceipt',{changes:changes, widget:this});
+                    printer.print(receipt);
+                    return delay(100);
+                });
             }
         },
         hasChangesToPrint: function(){